Add layer group support introduced in Gimp 2.8.
Changes introduced by this commit:

 - Update xcf-private.h with the file from the head of git's branch
   named gimp-2-8 (commit acebb4ad205901a9e8029a66ed5bca62cf13bc96).

 - Increase supported version number from 2 to 3.

 - Add the cases for PROP_GROUP_ITEM and PROP_ITEM_PATH in getBasicXcfInfo().

 - Add the function printLayerPath() to print the groups containing a given
   layer.

 - Add an option --path-separator (string) to set the string to use to
   separate the groups in the paths.

The output is unchanged for files whose version is lower than 3.
Otherwise, the name of the layer now contains its path in the image and
the marker "group" is added for the group layers in the fourth field.

For example, here is the output of a file containing a single layer
named "a", then one group named "G 1", which contains two layers named
"b 1" and "c 2". The path separator is "|" by default:

  Version 3, 10x10 RGB color, 4 layers, compressed RLE
  + 10x10+0+0 RGB-alpha Normal |a
  + 10x10+0+0 RGB-alpha Normal/group |G 1
  + 10x10+0+0 RGB-alpha Normal |G 1|b 1
  + 10x10+0+0 RGB-alpha Normal |G 1|c 2