Starting the migration to CMake build system in order to easily support Windows systems.
The configure stuff is now removed and the configuration is done via CMake.

The command line arguments where previously handled via some kind of magic,
where a file was parsed to extract the various argument formats and the code to
handle them (if and only if the output of the parser is included at the right
place in the source).

I have started the task to put the command line arguments in the code instead
of the parsed file. The help output is not handled yet and I have still to list
the commands used in each program. These changes add quite a lot of
uninteresting code but argument handling should become understandable.